Exponentially back off if we can't connect to Redis
Microservices shouldn't crash if something they depend on isn't available when
they start up.  Instead, this will retry as many times as it needs to to
connect to Redis with random intervals between the attempts.

When we implement healthchecks we'll make sure that the service does not appear
healthy unless it has an active redis connection.