/select-identity-provider: Check if IdP is enabled
This adds a call out to config to check if the IdP we're being asked to select
is actually enabled for this service.

There's a bit of grossness here in that the config service has two different
endpoints to call depending on whether the user is signing in or registering,
but there's no way around that without changing other services, which is out of
scope.

The base class for the tests now resets the redis cluster and the wiremocked
endpoints so that the tests are independent of each other.