Implement /sign-in-process-details
This has to call out to config to determine whether the transaction supports
eidas (this used to be cached in session, but that feels like a confusing
optimisation to me).

I've extended the HTTP timeouts to 1 second in the test jersey client becasue
sometimes the tests would fail due to wiremock taking too long to respond.