Implement /idp-authn-request-from-hub statefullness
This is a pretty hacky, but pretty simple approach. Instead of loading the
state machine up with knowledge of session information we're just putting the
required information into the session directly in the controller. Going
forwards we'll want to change this, but in the spirit of keeping things minimal
this is good enough for now.