Refactor: Pull out handlers for dealing with known situations
There are a few endpoints in policy where the pattern is basically:

- Pass the response off to saml engine
- Work out what to do given the type of response we got and the config
- Perform some side effects and transition the state

The handlers I'm pulling out just do the last bit. The handlers
shouldn't need to contain much in the way of logic.

In general I think a good pattern is to call the state transition at the
beginning of the method to check it's a legal transition, but not to
write it to Redis until the end of the method. This ensures that the
side effects don't happen unless the desired state transition is legal,
but if the side effects fail for some reason we don't end up setting an
incorrect state (although it's not at all clear from the original Policy
what to do to the state in the case where some of the side effects
fail).